Moved VecUtil and Interpolator into package `infovis.util'.
